+++ apps/core-lib-dev/src/app/app.element.ts
@@ -1,13 +1,18 @@
 import { css, LitElement } from 'lit'
+import { property, state } from 'lit/decorators.js'
+import { choose } from 'lit/directives/choose.js'
+
+import { html } from '@sebgroup/green-core/scoping'
+import { registerTransitionalStyles } from '@sebgroup/green-core/transitional-styles'
+import { gdsInitLocalization } from '@sebgroup/green-core/localization'
+import '@sebgroup/green-core/components/icon/icons/flag.js'
+import '@sebgroup/green-core/components/segmented-control/index.js'
+
 import './chlorophyll.scss'
 import './simple-values.element'
 import './complex-values.element'
-import './custom-localization.element'
-import './segmented-control.element'
-import { html } from '@sebgroup/green-core'
-import { registerTransitionalStyles } from '@sebgroup/green-core/transitional-styles'
 
-//import { setLocale } from '@sebgroup/green-core/localization'
+const { setLocale, getLocale } = gdsInitLocalization()
 
 export class AppElement extends LitElement {
   static styles = css`
@@ -15,21 +20,74 @@ export class AppElement extends LitElement {
       --gds-test-color: red;
     }
   `
+  // No need for a shadow root here
+  protected createRenderRoot() {
+    return this
+  }
+
+  @property()
+  accessor popoverOpen: boolean = false
+
+  @state()
+  accessor lang = 'sv'
+
+  @state()
+  accessor currentView = 'simple-values'
 
   connectedCallback() {
     super.connectedCallback()
-    //registerTransitionalStyles()
-    //setLocale('sv')
+    registerTransitionalStyles()
+    this.setLang(getLocale())
+  }
+
+  setLang(lang: string) {
+    this.lang = lang
+    setLocale(lang)
   }
 
   render() {
     return html`
-      <gds-theme color-scheme="dark">
-        <div class="wrapper">
-          <h1 class="mb-5" style="color: var(--gds-test-color)">
-            Green Core test app
-          </h1>
-          <slot></slot>
+      <gds-theme color-scheme="light">
+        <div class="container">
+          <div style="display: flex; justify-content: space-between;">
+            <h1 class="mb-5">Green Core test app</h1>
+            <gds-context-menu>
+              <span slot="trigger"
+                >Lang: ${this.lang} <gds-icon-flag></gds-icon-flag
+              ></span>
+              <gds-menu-item @click=${() => this.setLang('sv')}
+                >SV</gds-menu-item
+              >
+              <gds-menu-item @click=${() => this.setLang('en')}
+                >EN</gds-menu-item
+              >
+            </gds-context-menu>
+          </div>
+
+          <gds-segmented-control
+            .value=${this.currentView}
+            @change=${(e: CustomEvent) =>
+              (this.currentView = (e.target as any).value)}
+            style="margin-bottom: 1rem;"
+          >
+            <gds-segment value="simple-values">Form: Simple values</gds-segment>
+            <gds-segment value="complex-values"
+              >Form: Complex values</gds-segment
+            >
+          </gds-segmented-control>
+          <div class="card">
+            ${choose(
+              this.currentView,
+              [
+                ['simple-values', () => html`<simple-values></simple-values>`],
+                [
+                  'complex-values',
+                  () => html`<complex-values></complex-values>`,
+                ],
+              ],
+              () => html`No view selected`,
+            )}
+          </div>
         </div>
       </gds-theme>
     `