+++ libs/core/src/components/grid/grid.ts
@@ -11,8 +11,6 @@ import { tokens } from '../../tokens.style'
 
 import GridCSS from './grid.style.css'
 
-type GridSizes = 'none' | 'xs' | 's' | 'm' | 'l' | 'xl' | '2xl' | '3xl'
-
 const BreakpointPattern =
   /(?<l>l:([a-z0-9]+))?\s*(?<m>m:([a-z0-9]+))?\s*(?<s>s:([a-z0-9]+))?/
 
@@ -57,7 +55,7 @@ export class GdsGrid extends GdsElement {
    * ```
    */
   @property({ attribute: 'gap', type: String })
-  gap?: GridSizes
+  gap?: string
 
   /**
    * Defines the gap size between grid items in vertical axis. Accepts a single value for all breakpoints or a "l:desktop m:tablet s:mobile" format. Sizes can be 'none', 'xs', 's', 'm', 'l', 'xl', '2xl', '3xl'.
@@ -68,7 +66,7 @@ export class GdsGrid extends GdsElement {
    * ```
    */
   @property({ attribute: 'row-gap', type: String })
-  rowGap?: GridSizes
+  rowGap?: string
 
   /**
    * Defines the padding size around the grid. Accepts a single value for all breakpoints or a "l:desktop m:tablet s:mobile" format. Sizes can be 'none', 'xs', 's', 'm', 'l', 'xl', '2xl', '3xl'.
@@ -79,7 +77,7 @@ export class GdsGrid extends GdsElement {
    * ```
    */
   @property({ attribute: 'padding', type: String })
-  padding?: GridSizes
+  padding?: string
 
   /**
    * Defines the minimum column width in pixels. Accepts a single value for all breakpoints or a "l:desktop m:tablet s:mobile" format. If set, the grid adjusts column size based on content and available width, even without other attributes.
@@ -90,7 +88,7 @@ export class GdsGrid extends GdsElement {
    * ```
    */
   @property({ attribute: 'auto-columns', type: String })
-  autoColumns?: GridSizes
+  autoColumns?: string
 
   /**
    * Lifecycle method called when the element is connected to the DOM.
@@ -172,7 +170,7 @@ export class GdsGrid extends GdsElement {
       }
     }
 
-    if (propertyName === 'auto-columns') {
+    if (propertyName === 'autoColumns') {
       if (this.autoColumns && !this.autoColumns.includes(' ')) {
         desktop = tablet = mobile = `${this.autoColumns}px`
       } else {
@@ -264,7 +262,7 @@ export class GdsGrid extends GdsElement {
    */
   @watch('autoColumns')
   private _updateAutoColumnsVariables() {
-    this._updateCSSVariables('auto-columns', 'col-width')
+    this._updateCSSVariables('autoColumns', 'col-width')
   }
 
   render() {