+++ libs/angular/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 4.1.0
+
+### Minor Changes
+
+- 5a9e3ab: Modal animations
+
+### Patch Changes
+
+- Updated dependencies [a8b556f]
+- Updated dependencies [096b662]
+- Updated dependencies [5f224d0]
+- Updated dependencies [1652e33]
+  - @sebgroup/green-core@1.22.0
+
 ## 4.0.0
 
 ### Major Changes